[ refactor ] Reservation Message DTO 구조 개선 및 코드 리팩토링 ( #17 )
- ReservationCreateEventDTOV1 -> CreateReservationMessageDTOV1로 DTO 명 변경

- 이전 구조에서 `Message`, `User`, `Restaurant`, `Reservation`을 포함하는 방식에서, 필요한 필드만 포함하는 간결한 구조로 변경

- `of()` 메서드를 통해 객체 생성 로직을 캡슐화
